* amandad-src/amandad.c: Remove duplicate setsockopt(,,SO_REUSEADDR). * common-src/bsd-security.c: Use int for setsockopt. * common-src/stream.c (try_socksize): Use int for setsockopt.
git-svn-id: https://amanda.svn.sourceforge.net/svnroot/amanda/amanda/trunk@263 a8d146d6-cc15-0410-8900-af154a0219e0

+++ common-src/bsd-security.c
@@ -365,7 +365,7 @@ bsd_stream_client(
     struct sec_stream *bs = NULL;
     struct sec_handle *bh = h;
 #ifdef DUMPER_SOCKET_BUFFERING
-    size_t rcvbuf = SIZEOF(bs->databuf) * 2;
+    int rcvbuf = SIZEOF(bs->databuf) * 2;
 #endif
 
     assert(bh != NULL);
